[Attributor] Allow persistent arguments to replace values
When we determine if a value is dead, we checked if it was replaced by a
constant. We avoided other replacement values because without the use,
which is only added at the end, they might be considered dead and
eliminated themselves. This cannot happen for persistent arguments,
hence we can allow them in addition to constants.